+++ src/responses.rs
@@ -1,9 +1,15 @@
-use std::sync::mpsc::{Receiver, Sender};
+use crate::widgets::{Widget, WidgetConfig};
+use std::{
+    cell::RefCell,
+    rc::Rc,
+    sync::mpsc::{Receiver, Sender},
+};
 /// This is the struct that gets returned when a widget is added to the context.
 pub struct Response<R> {
     ///This is used to comunicate with the widget
     pub channel: R,
     pub(crate) _id: WidgetId,
+    pub(crate) _layer_id: LayerId,
 }
 
 pub(crate) type LayerNummerId = u64;
@@ -16,33 +22,89 @@ pub(crate) type WidgetChannelReceiver =
     Receiver<((LayerNummerId, WidgetNummerId), WidgetInstruction)>;
 pub(crate) enum LayerInstructions {
     Drop,
-    SetIsActive(bool),
+    AddWidget(Box<dyn Widget + 'static>, WidgetNummerId),
 }
 
 pub(crate) enum WidgetInstruction {
     Drop,
 }
 
-///Used to keep set to which layer a widget belongs to.
-///Dropping this causes every widget in this layer to be removed.
+///Used to create widgets at a layer.
+///Once this and every widget on this layer are dropped, so is the internal layer.
+#[derive(Clone)]
 pub struct LayerId {
-    pub(crate) id: LayerNummerId,
-    channel: LayerChannelSender,
+    layer: Rc<InternalLayerId>,
+    is_active: Rc<RefCell<bool>>,
+    widget_id: Rc<RefCell<WidgetNummerId>>,
+    widget_channel: WidgetChannelSender,
 }
+
 impl LayerId {
-    pub(crate) fn new(id: LayerNummerId, channel: LayerChannelSender) -> Self {
-        Self { id, channel }
+    pub(crate) fn new(
+        id: LayerNummerId,
+        channel: LayerChannelSender,
+        is_active: Rc<RefCell<bool>>,
+        widget_id: Rc<RefCell<WidgetNummerId>>,
+        widget_channel: WidgetChannelSender,
+    ) -> Self {
+        let layer = Rc::new(InternalLayerId::new(id, channel));
+        Self {
+            layer,
+            is_active,
+            widget_id,
+            widget_channel,
+        }
+    }
+
+    ///Adds a widget configuration to the layer that this id represents.
+    ///Returns a channel to comunicate with the widget.
+    pub fn add_widget<ReturnChannel, W: Widget + 'static>(
+        &mut self,
+        widget_config: impl WidgetConfig<ReturnChannel, W>,
+    ) -> Response<ReturnChannel> {
+        let (widget, res) = widget_config.to_widget();
+        let mut widget_id = self.widget_id.borrow_mut();
+        *widget_id += 1;
+        let x = Response {
+            _layer_id: self.clone(),
+            channel: res,
+            _id: WidgetId::new(self.id(), *widget_id, self.widget_channel.clone()),
+        };
+        self.layer.as_ref().add_widget(widget, *widget_id);
+        x
+    }
+
+    pub(crate) fn id(&self) -> LayerNummerId {
+        self.layer.as_ref().id
     }
 
     ///Set a layer to active or inactive.
     ///Layers that are inactive won't be rendered or receive updates.
     pub fn set_is_active(&self, is_active: bool) {
-        let _ = self
-            .channel
-            .send((self.id, LayerInstructions::SetIsActive(is_active)));
+        self.is_active.replace(is_active);
+    }
+
+    ///Get if the layer is active or not.
+    pub fn get_active(&self) -> bool {
+        *self.is_active.borrow()
     }
 }
-impl Drop for LayerId {
+pub(crate) struct InternalLayerId {
+    pub(crate) id: LayerNummerId,
+    channel: LayerChannelSender,
+}
+
+impl InternalLayerId {
+    pub(crate) fn new(id: LayerNummerId, channel: LayerChannelSender) -> Self {
+        Self { id, channel }
+    }
+    pub(crate) fn add_widget(&self, widget: impl Widget + 'static, id: WidgetNummerId) {
+        let _ = self
+            .channel
+            .send((self.id, LayerInstructions::AddWidget(Box::new(widget), id)));
+    }
+}
+impl Drop for InternalLayerId {
     fn drop(&mut self) {
         let _ = self.channel.send((self.id, LayerInstructions::Drop));
     }